Refactor find actions
