+++ src/actions/complete-file-upload.js
@@ -3,10 +3,8 @@ const { get } = require('lodash');
 module.exports = function({ request, uploadCompleteRoute }) {
   /**
    * Marks the file upload as completed.
-   *
    * @param   {Object} transferOrBoard Transfer or board item.
    * @param   {Object} file            File containing information about number of parts, upload url, etc.
-   *
    * @returns {Promise}
    */
   return function completeFileUpload(transferOrBoard, file) {
